lsp: move Backend and Rust to modules
